gui: remove warning

+++ gui/src/Gui_draw.cpp
@@ -173,9 +173,9 @@ void Gui::draw_map_tile(int i, int j, int tile)
 void Gui::draw_map_half_tile(int i, int j, int tile)
 {
     int x = tile - '0';
-    if (x == 11 && (i % 5 == _waves % 5))
+    if (x == 11 && ((size_t)(i % 5) == _waves % 5))
         x = 18;
-    if (x == 11 && (j + i) % 5 == _waves % 5)
+    if (x == 11 && (size_t)((j + i) % 5) == _waves % 5)
         x = 13;
     int y = 0;
     while (x >= 10)
@@ -208,7 +208,7 @@ void Gui::draw_map_half_tile(int i, int j, int tile)
 
 void Gui::draw_map(void)
 {
-    for (int i = 0; i < _map_decor.size(); i++)
+    for (int i = 0; (size_t)i < _map_decor.size(); i++)
     {
         for (int j = _map_decor[i].size() - 1; j >= 0; j--)
         {
